BirdHouse v0.5
Make compatible with Titanium Developer on Mac for both Android
and iPhone. Handle authorization process when other than 'allow'
is clicked. Alter tweet UI for Android and iPhone. Add callback
functionality.

+++ birdhouse.js
@@ -5,13 +5,21 @@
 // authenticating and sending API calls to Twitter.
 //
 // Author: Joseph D. Purcell, iEntry Inc.
-// Version: 0.4
+// Version: 0.5
 // Modified: April 2011
 // --------------------------------------------------------
 
 // INCLUDES
-Ti.include('oauth.js');
-Ti.include('sha1.js');
+// iphone requires complete path
+if (Ti.Platform.osname=='iphone') {
+	Ti.include('lib/oauth.js');
+	Ti.include('lib/sha1.js');
+}
+// android will search the current directory for files
+else {
+	Ti.include('oauth.js');
+	Ti.include('sha1.js');
+}
 
 // THE CLASS
 function BirdHouse(params) {
@@ -50,18 +58,18 @@ function BirdHouse(params) {
 	// In Parameters:
 	//	url (String) - the url to send the message to
 	//	method (String) - 'POST' or 'GET'
-	//	parameters (String) - parameters to add to the
+	//	params (String) - parameters to add to the
 	//	  message in URL form, i.e. var1=2&var2=3
 	//
 	// Returns:
 	//	message (Array) - the message parameters to send
 	//	  to Twitter
 	// --------------------------------------------------------
-	function set_message(url, method, parameters) {
+	function set_message(url, method, params) {
 		var message = {
 			action: url,
-			method: (method) ? method : 'POST',
-			parameters: (parameters!=null) ? OAuth.decodeForm(parameters) : []
+			method: (method=='GET') ? method : 'POST',
+			parameters: (params!=null) ? OAuth.decodeForm(params) : []
 		};
 		message.parameters.push(['oauth_consumer_key', cfg.oauth_consumer_key]);
 		message.parameters.push(['oauth_signature_method', cfg.oauth_signature_method]);
@@ -123,40 +131,90 @@ function BirdHouse(params) {
 		var request_token = "";
 		var url_base = "";
 		var params = "";
+		var compareUrl = ""; // used for iOS
 		var win = Ti.UI.createWindow({
 			top: 0,
 			modal: true,
 			fullscreen: true
 		});
+		var init = true;
+		var allow = false;
+		var title = '';
 
 		// add the webview to the window and open the window
 		win.add(webView);
 		win.open();
 
-		// on url change, see if 'oauth_verifier' is in the url
-		webView.addEventListener('load',function(){
-			params = "";
-			var parts = (webView.url).replace(/[?&]+([^=&]+)=([^&]*)/gi, function(m,key,value) {
-				params = params + m;
+		webView.addEventListener('load',function(e){
+			title = webView.evalJS("document.title");
+			// the first time load, ignore
+			if (init) {
+				init = false
+			}
+			// the second load they have clicked allow or deny
+			else {
+				// listen for the tokenses if "allow" was clicked
+				if (allow) {
+					// success!
+					params = "";
+					var parts = (e.url).replace(/[?&]+([^=&]+)=([^&]*)/gi, function(m,key,value) {
+						params = params + m;
 
-				if (key=='oauth_verifier') {
-					cfg.request_verifier = value;
+						if (key=='oauth_verifier') {
+							cfg.request_verifier = value;
+						}
+					});
+
+					if (cfg.request_verifier!="") {
+						// my attempt at making sure the stupid webview dies
+						webView.stopLoading();
+						win.remove(webView);
+						win.close();
+
+						get_access_token(callback);
+
+						return true; // we are done here
+					}
 				}
-			});
+				// find out what the user is clicking on
+				else {
 
-			// success!
-			if (cfg.request_verifier!="") {
-				// my attempt at making sure the stupid webview dies
-				webView.stopLoading();
-				win.remove(webView);
-				win.close();
+					// user is creating a new account (goes to a suspended page if clicked second time)
+					// just kill the auth process and say "go do this somewhere else"
+					if ((e.url).search('account')!=-1 || (e.url).search('suspended')!=-1) {
+						webView.stopLoading();
+						win.remove(webView);
+						win.close();
 
-				get_access_token(callback);
+						if(typeof(callback)=='function'){
+							callback(false);
+						}
 
-				return true; // we are done here
+						return false;
+					}
+					// access denied was clicked
+					else if ((webView.evalJS("document.title")).search('Redirect')==-1) {
+						webView.stopLoading();
+						win.remove(webView);
+						win.close();
+
+						if(typeof(callback)=='function'){
+							callback(false);
+						}
+
+						return false;
+					}
+					// allow access was clicked
+					else {
+						// note that we can skip the current page, because the page
+						// is a page on Twitter with a javscript (or PHP) redirect
+						// when webView load fires again, allow will be true
+						// and we will extract the tokenses
+						allow = true;
+					}
+				}
 			}
 		});
-
 	}
 
 	// --------------------------------------------------------
@@ -175,7 +233,6 @@ function BirdHouse(params) {
 
 		api(url,'POST','oauth_token='+cfg.request_token+'&oauth_verifier='+cfg.request_verifier,function(resp){
 			if (resp!=false) {
-				// SUCCESS
 				var responseParams = OAuth.getParameterMap(resp);
 				cfg.access_token = responseParams['oauth_token'];
 				cfg.access_token_secret = responseParams['oauth_token_secret'];
@@ -192,7 +249,6 @@ function BirdHouse(params) {
 					callback(true);
 				}
 			} else {
-				// FAIL
 				// execute the callback function
 				if(typeof(callback)=='function'){
 					callback(false);
@@ -261,7 +317,6 @@ function BirdHouse(params) {
 			access_token_secret: cfg.access_token_secret
 		};
 		file.write(JSON.stringify(config));
-
 	}
 
 	// --------------------------------------------------------
@@ -283,17 +338,31 @@ function BirdHouse(params) {
 	//
 	// Notes:
 	//	- the setUrlParams and setHeader should only need
-	//	  to be set whenever getting request and access
-	//	  tokens; values 'true' and 'false' respectively
+	//	  to be set whenever getting request tokens; values
+	//	  should be 'true' and 'false' respectively
+	//	- take advantage of the callback function
 	//
 	// Returns: false on failure and the responseText on
 	//   success.
 	// --------------------------------------------------------
 	function api(url, method, params, callback, auth, setUrlParams, setHeader) {
+		var finalUrl = '';
+
 		// authorize user if not authorized, and call this in the callback
-		if (!authorized && (typeof(auth)=='undefined' || auth==true)) {
-			authorize(function(){
-				api(url,method,params,callback,auth);
+		if (!authorized && (typeof(auth)=='undefined' || auth===true)) {
+			authorize(function(retval){
+				if (!retval) {
+					// we have returned from authorizing, but auth is false, so return false & dont execute API
+					// execute the callback function
+					if (typeof(callback)=='function') {
+						callback(false);
+					}
+
+					return false;
+				} else {
+					// fn-api: we have returned from authorizing & we are authorized!
+					api(url,method,params,callback,auth);
+				}
 			});
 		}
 		// user is authorized so execute API
@@ -309,25 +378,26 @@ function BirdHouse(params) {
 			// VARIABLES
 			var initparams = params;
 
+			// set params
 			if (params!=null) {
 				params = params + "&";
 			}
-
 			if (cfg.access_token!='') {
 				params = params + "oauth_token="+cfg.access_token;
 			}
-			var message = set_message(url, method, params);
-			//var message = createMessage(url, method, params);
 
+			// set & sign the message
+			var message = set_message(url, method, params);
 			OAuth.SignatureMethod.sign(message, accessor);
 
+			// set the URL to send request to
 			// if we are getting request tokens, all params have to be set in URL
 			if (typeof(setUrlParams)!='undefined' && setUrlParams==true) {
-				var finalUrl = OAuth.addToURL(message.action, message.parameters);
+				finalUrl = OAuth.addToURL(message.action, message.parameters);
 			}
 			// for all other requests only custom params need set in the URL
 			else {
-				var finalUrl = OAuth.addToURL(message.action, initparams);
+				finalUrl = OAuth.addToURL(message.action, initparams);
 			}
 
 			var XHR = Ti.Network.createHTTPClient();
@@ -335,7 +405,7 @@ function BirdHouse(params) {
 			// on success, grab the request token
 			XHR.onload = function() {
 				// execute the callback function
-				if(typeof(callback)=='function'){
+				if (typeof(callback)=='function') {
 					callback(XHR.responseText);
 				}
 
@@ -344,6 +414,7 @@ function BirdHouse(params) {
 
 			// on error, show message
 			XHR.onerror = function(e) {
+				// access token and token secret are wrong
 				return false;
 			}
 			
@@ -352,17 +423,17 @@ function BirdHouse(params) {
 			// if we are getting request tokens do not set the HTML header
 			if (typeof(setHeader)=='undefined' || setHeader==true) {
 				var init = true;
-				var auth = "OAuth ";
+				var header = "OAuth ";
 				for (var i=0; i<message.parameters.length; i++) {
 					if (init) {
 						init = false;
 					} else {
-						auth = auth + ",";
+						header = header + ",";
 					}
-					auth = auth + message.parameters[i][0] + '="' + escape(message.parameters[i][1]) + '"';
+					header = header + message.parameters[i][0] + '="' + escape(message.parameters[i][1]) + '"';
 				}
 
-				XHR.setRequestHeader("Authorization", auth);
+				XHR.setRequestHeader("Authorization", header);
 			}
 			
 			XHR.send();
@@ -379,63 +450,128 @@ function BirdHouse(params) {
 	//
 	// In Parameters:
 	//	text (String) - the default text for the text area
+	//	callback (Function) - function to call after XHR
 	// --------------------------------------------------------
-	function tweet(text) {
+	function tweet(text,callback) {
 		// VALIDATE INPUT
+		// just in case someone only wants to send a callback
+		if (typeof(text)=='function' && typeof(callback)=='undefined') {
+			callback = text;
+			text = '';
+		}
 		if (typeof(text)=='undefined') {
 			text = '';
 		}
+
 		var obj = this;
 		obj.mytweet = text;
 		if (authorized === false) {
 			authorize(function(resp){
 				if (resp) {
 					obj.tweet(obj.mytweet);
+
+					return true;
+				} else {
+					// execute the callback function
+					if (typeof(callback)=='function') {
+						callback(false);
+					}
+
+					return false;
 				}
 			});
 		} else {
 			var chars = (typeof(text)!='undefined' && text!=null)?text.length:0;
+
 			var winBG = Titanium.UI.createWindow({
 				backgroundColor:'#000',
 				opacity:0.60
 			});
-			var winTW = Titanium.UI.createWindow({
-				height:304,
-				top:10,
-				right:10,
-				left:10,
-				borderColor:'#224466',
-				borderWidth:3,
-				backgroundColor:'#559abb',
-				borderRadius:3.0
-			});
-			var tweet = Ti.UI.createTextArea({
-				value:text,
-				height:200,
-				top:14,
-				left:14,
-				right:14
-			});
-			var btnTW = Ti.UI.createButton({
-				title:'Tweet',
-				width:100,
-				top:222,
-				right:24
-			});
-			var btnCancel = Ti.UI.createButton({
-				title:'Cancel',
-				width:100,
-				top:222,
-				left:24
-			});
-			var charcount = Ti.UI.createLabel({
-				bottom:10,
-				right:14,
-				color:'#FFF',
-				text:parseInt((140-chars))+''
-			});
-			tweet.addEventListener('change',function() {
-				chars = (140-this.value.length);
+
+			// the UI window looks completely different on iPhone vs. Android
+			// iPhone UI
+			if (Ti.Platform.osname=='iphone') {
+				var winTW = Titanium.UI.createWindow({
+					height:((Ti.Platform.displayCaps.platformHeight*0.5)-15), // half because the keyboard takes up half
+					width:(Ti.Platform.displayCaps.platformWidth-20),
+					top:10,
+					right:10,
+					left:10,
+					borderColor:'#224466',
+					borderWidth:3,
+					backgroundColor:'#559abb',
+					borderRadius:3.0
+				});
+				var tweet = Ti.UI.createTextArea({
+					value:text,
+					height:((Ti.Platform.displayCaps.platformHeight*0.5)-100),
+					width:(Ti.Platform.displayCaps.platformWidth-48),
+					font:{fontSize:16},
+					top:14,
+					left:14,
+					right:14
+				});
+				var btnTW = Ti.UI.createButton({
+					title:'Tweet!',
+					width:100,
+					height:30,
+					top:((Ti.Platform.displayCaps.platformHeight*0.5)-75),
+					right:24
+				});
+				var btnCancel = Ti.UI.createButton({
+					title:'Cancel',
+					width:100,
+					height:30,
+					top:((Ti.Platform.displayCaps.platformHeight*0.5)-75),
+					left:24
+				});
+				var charcount = Ti.UI.createLabel({
+					top:((Ti.Platform.displayCaps.platformHeight*0.5)-55),
+					left:(Ti.Platform.displayCaps.platformWidth-60), // 30 px from right side,
+					color:'#FFF',
+					text:(parseInt((140-chars))+'')
+				});
+			}
+			// Android UI
+			else {
+				var winTW = Titanium.UI.createWindow({
+					height:264,
+					top:10,
+					right:10,
+					left:10,
+					borderColor:'#224466',
+					borderWidth:3,
+					backgroundColor:'#559abb',
+					borderRadius:3.0
+				});
+				var tweet = Ti.UI.createTextArea({
+					value:text,
+					height:160,
+					top:14,
+					left:14,
+					right:14
+				});
+				var btnTW = Ti.UI.createButton({
+					title:'Tweet',
+					width:100,
+					top:182,
+					right:24
+				});
+				var btnCancel = Ti.UI.createButton({
+					title:'Cancel',
+					width:100,
+					top:182,
+					left:24
+				});
+				var charcount = Ti.UI.createLabel({
+					bottom:10,
+					right:14,
+					color:'#FFF',
+					text:(parseInt((140-chars))+'')
+				});
+			}
+			tweet.addEventListener('change',function(e) {
+				chars = (140-e.value.length);
 				if (chars<11) {
 					if (charcount.color!='#D40D12') {
 						charcount.color = '#D40D12';
@@ -459,12 +595,35 @@ function BirdHouse(params) {
 						buttonNames: ['OK']
 					});
 					alertDialog.message = "Tweet failed to send!";
+
+					// execute the callback function
+					if (typeof(callback)=='function') {
+						callback(false);
+					}
+
+					return false;
 				} else {
+					// hide the keyboard on Android because it doesn't automatically
+					if (Ti.Platform.osname=='android') {
+						Titanium.UI.Android.hideSoftKeyboard();
+					}
+
+					// execute the callback function
+					if (typeof(callback)=='function') {
+						callback(false);
+					}
+
 					winBG.close();
 					winTW.close();
+
+					return true;
 				}
 			});
 			btnCancel.addEventListener('click',function() {
+				// hide the keyboard on Android because it doesn't automatically
+				if (Ti.Platform.osname=='android') {
+					Titanium.UI.Android.hideSoftKeyboard();
+				}
 				winBG.close();
 				winTW.close();
 			});
@@ -489,10 +648,8 @@ function BirdHouse(params) {
 	function send_tweet(params) {
 		api('http://api.twitter.com/1/statuses/update.json','POST',params,function(resp){
 			if (resp!=false) {
-				// SUCCESS
 				return true;
 			} else {
-				// FAIL
 				return false;
 			}
 		});
@@ -501,14 +658,36 @@ function BirdHouse(params) {
 	// --------------------------------------------------------
 	// get_tweets
 	//
-	// Makes a TWitter API call to get tweets.
+	// Makes a TWitter API call to get tweets from the friends
+	// timeline.
 	//
 	// In Parameters:
 	//	params (String) - the string of optional and
 	//	  required parameters in url form
+	//	callback (Function) - function to call after XHR
 	// --------------------------------------------------------
-	function get_tweets(params) {
-		return api("https://api.twitter.com/1/statuses/friends_timeline.json","GET",params);
+	function get_tweets(params,callback) {
+		// just in case someone only wants to send a callback
+		if (typeof(params)=='function' && typeof(callback)=='undefined') {
+			callback = params;
+			params = '';
+		}
+
+		api("https://api.twitter.com/1/statuses/friends_timeline.json","GET",params,function(tweets){
+			try {
+				// parse the JSON, because there isn't a case where you want a string instead of obj returned
+				tweets = JSON.parse(tweets);
+			} catch (e) {
+				tweets = false;
+			}
+
+			// execute the callback function
+			if(typeof(callback)=='function'){
+				callback(tweets);
+			}
+
+			return tweets;
+		})
 	}
 
 	// --------------------------------------------------------
@@ -522,10 +701,29 @@ function BirdHouse(params) {
 	// In Parameters:
 	//	callback (Function) - a function to call after
 	//	  the user has been authorized; note that it won't
-	//	  be executed until get_access_token()
+	//	  be executed until get_access_token(), unless we
+	//	  are already authorized, and we will call it here
+	//
+	// Returns: true if the user is authorized, else false
 	// --------------------------------------------------------
 	function authorize(callback) {
-		get_request_token(callback);
+		if (!authorized) {
+			get_request_token(callback); // get_request_token or a function it calls will call callback
+		} else {
+			var alertDialog = Titanium.UI.createAlertDialog({
+				title: 'Error Message',
+				message: 'You are already authorized!',
+				buttonNames: ['Schweet!']
+			});
+			alertDialog.show();
+
+			// execute the callback function
+			if(typeof(callback)=='function'){
+				callback(authorized);
+			}
+		}
+
+		return authorized;
 	}
 
 	// --------------------------------------------------------
@@ -536,14 +734,45 @@ function BirdHouse(params) {
 	// load_access_token() which should return false since
 	// we deleted the file, thus resulting in a deauthroized
 	// state.
+	//
+	// In Parameters:
+	//	callback (Function) - function to call after
+	//	  user is deauthorized
+	//
+	// Returns: true if the user is deauthorized, else false
 	// --------------------------------------------------------
-	function deauthorize() {
-		var file = Ti.Filesystem.getFile(Ti.Filesystem.applicationDataDirectory, 'twitter.config');
-		file.deleteFile();
-		authorized = load_access_token();
-		accessor.tokenSecret = "";
-		cfg.access_token = "";
-		cfg.access_token_secret = "";
+	function deauthorize(callback) {
+		if (authorized) {
+			// delete the config file
+			var file = Ti.Filesystem.getFile(Ti.Filesystem.applicationDataDirectory, 'twitter.config');
+			file.deleteFile();
+			// check to make sure it was deleted by re-setting authorized
+			authorized = load_access_token();
+			// reset config
+			accessor.tokenSecret = "";
+			cfg.access_token = "";
+			cfg.access_token_secret = "";
+			cfg.request_verifier = "";
+
+			// execute the callback function
+			if(typeof(callback)=='function'){
+				callback(!authorized);
+			}
+		} else {
+			var alertDialog = Titanium.UI.createAlertDialog({
+				title: 'Error Message',
+				message: 'You are already deauthorized!',
+				buttonNames: ['Oh, OK']
+			});
+			alertDialog.show();
+
+			// execute the callback function
+			if(typeof(callback)=='function'){
+				callback(!authorized);
+			}
+		}
+
+		return !authorized;
 	}
 
 	// --------------------------------------------------------
